server/requirements.txt
asgiref==3.3.1
chardet==4.0.0
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
gunicorn==20.1.0
networkx=2.8=pyhd8ed1ab_0
osmnx=1.1.2=pyhd8ed1ab_0
psycopg2-binary==2.8.6
pyjwt==2.1.0
python-dotenv==0.15.0
pytz==2021.1
sqlparse==0.4.1